requirements.txt
flask>=3.0
flask-cors>=4.0
mysql-connector-python>=8.0
python-dotenv>=1.0
pandas>=2.0
numpy>=1.24
scipy>=1.10